requirements.txt
Django>=4.0,<5.0
psycopg>=2.8
environs>10.0.0
djangorestframework==3.14.0